refactor: split WalletModal into focused sub-components
Extracts 10 files from the 2129 LOC monolith (now 344 LOC, 84% reduction):
- wallet-modal/utils.ts (shared types + constants + helpers)
- wallet-modal/WalletAddressCard.tsx
- wallet-modal/WalletReceiveView.tsx
- wallet-modal/WalletSendForm.tsx
- wallet-modal/WalletFundMenu.tsx
- wallet-modal/WalletTokenList.tsx
- wallet-modal/WalletAmountStep.tsx
- wallet-modal/CountdownBadge.tsx
- wallet-modal/WalletConfirmStep.tsx
- wallet-modal/WalletSuccessStep.tsx
